bug 1462749: move graphics devices to django admin
