+++ webapp-django/crashstats/manage/admin.py
@@ -3,13 +3,19 @@ import hashlib
 import requests
 from six.moves.urllib.parse import urlparse
 
+from django import http
 from django.conf import settings
+from django.contrib import messages
 from django.core.cache import cache
 from django.db import connection
-from django.shortcuts import render
+from django.shortcuts import redirect, render
 
 from crashstats.manage.decorators import superuser_required
+from crashstats.manage import forms
+from crashstats.manage import utils
 from crashstats.supersearch.models import SuperSearchMissingFields
+from crashstats.crashstats.models import GraphicsDevices
+from crashstats.crashstats.utils import json_view
 
 
 @superuser_required
@@ -127,3 +133,67 @@ def supersearch_fields_missing(request):
     context['title'] = 'Super search missing fields'
 
     return render(request, 'admin/supersearch_fields_missing.html', context)
+
+
+@superuser_required
+def graphics_devices(request):
+    context = {}
+    form = forms.GraphicsDeviceForm()
+    upload_form = forms.GraphicsDeviceUploadForm()
+
+    if request.method == 'POST' and 'file' in request.FILES:
+        upload_form = forms.GraphicsDeviceUploadForm(
+            request.POST,
+            request.FILES
+        )
+        if upload_form.is_valid():
+            if upload_form.cleaned_data['database'] == 'pcidatabase.com':
+                function = utils.pcidatabase__parse_graphics_devices_iterable
+            else:
+                function = utils.pci_ids__parse_graphics_devices_iterable
+
+            payload = list(function(upload_form.cleaned_data['file']))
+            api = GraphicsDevices()
+            result = api.post(data=payload)
+            messages.success(
+                request,
+                'Graphics device CSV upload successfully saved.'
+            )
+            return redirect('siteadmin:graphics_devices')
+
+    elif request.method == 'POST':
+        form = forms.GraphicsDeviceForm(request.POST)
+        if form.is_valid():
+            payload = [{
+                'vendor_hex': form.cleaned_data['vendor_hex'],
+                'adapter_hex': form.cleaned_data['adapter_hex'],
+                'vendor_name': form.cleaned_data['vendor_name'],
+                'adapter_name': form.cleaned_data['adapter_name'],
+            }]
+            api = GraphicsDevices()
+            result = api.post(data=payload)
+            if result:
+                messages.success(
+                    request,
+                    'Graphics device saved.'
+                )
+            return redirect('siteadmin:graphics_devices')
+
+    context['title'] = "Graphics devices"
+    context['form'] = form
+    context['upload_form'] = upload_form
+    return render(request, 'admin/graphics_devices.html', context)
+
+
+@json_view
+@superuser_required
+def graphics_devices_lookup(request):
+    form = forms.GraphicsDeviceLookupForm(request.GET)
+    if form.is_valid():
+        vendor_hex = form.cleaned_data['vendor_hex']
+        adapter_hex = form.cleaned_data['adapter_hex']
+        api = GraphicsDevices()
+        result = api.get(vendor_hex=vendor_hex, adapter_hex=adapter_hex)
+        return result
+    else:
+        return http.HttpResponseBadRequest(str(form.errors))
