+++ webapp-django/crashstats/manage/admin_urls.py
@@ -10,6 +10,12 @@ urlpatterns = [
     url('^crash-me-now/$',
         admin.crash_me_now,
         name='crash_me_now'),
+    url('^graphics-devices/$',
+        admin.graphics_devices,
+        name='graphics_devices'),
+    url('^graphics-devices/lookup/$',
+        admin.graphics_devices_lookup,
+        name='graphics_devices_lookup'),
     url('^sitestatus/$',
         admin.site_status,
         name='site_status'),
