+++ webapp-django/crashstats/manage/jinja2/manage/base.html
@@ -13,8 +13,6 @@
             <ul class="options">
                 <li><a href="{{ url('manage:home') }}"
                     {% if request.path_info == url('manage:home') %}class="selected"{% endif %}>Home</a></li>
-                <li><a href="{{ url('manage:graphics_devices') }}"
-                    {% if url( 'manage:graphics_devices') in request.path_info %}class="selected"{% endif %}>Graphics Devices</a></li>
             </ul>
         </nav>
     </div>
