+++ webapp-django/crashstats/manage/jinja2/manage/home.html
@@ -13,13 +13,6 @@
 
 <div class="panel">
     <div class="body">
-        <h3>Data</h3>
-        <ul>
-            <li><a href="{{ url('manage:graphics_devices') }}">
-                <span class="fa fa-desktop fa-5x"></span>
-                Graphics Devices
-            </a></li>
-        </ul>
     </div>
 </div>
