+++ webapp-django/crashstats/manage/templates/admin/socorro_admin_index.html
@@ -36,6 +36,13 @@
             </tr>
             <tr>
                 <th scope="row">
+                    <a href="{% url 'siteadmin:graphics_devices' %}">Graphics devices</a>
+                </th>
+                <td></td>
+                <td></td>
+            </tr>
+            <tr>
+                <th scope="row">
                     <a href="{% url 'siteadmin:site_status' %}">Site status</a>
                 </th>
                 <td></td>
