+++ webapp-django/crashstats/manage/urls.py
@@ -8,10 +8,4 @@ urlpatterns = [
     url('^$',
         views.home,
         name='home'),
-    url('^graphics-devices/$',
-        views.graphics_devices,
-        name='graphics_devices'),
-    url('^graphics-devices/lookup/$',
-        views.graphics_devices_lookup,
-        name='graphics_devices_lookup'),
 ]
