+++ webapp-django/crashstats/manage/views.py
@@ -1,82 +1,9 @@
-from builtins import str
+from django.shortcuts import render
 
-from django import http
-from django.contrib import messages
-from django.shortcuts import redirect, render
-
-from crashstats.crashstats.models import GraphicsDevices
-from crashstats.crashstats.utils import json_view
 from crashstats.manage.decorators import superuser_required
-
-from . import forms
-from . import utils
 
 
 @superuser_required
 def home(request, default_context=None):
     context = default_context or {}
     return render(request, 'manage/home.html', context)
-
-
-@superuser_required
-def graphics_devices(request):
-    context = {}
-    form = forms.GraphicsDeviceForm()
-    upload_form = forms.GraphicsDeviceUploadForm()
-
-    if request.method == 'POST' and 'file' in request.FILES:
-        upload_form = forms.GraphicsDeviceUploadForm(
-            request.POST,
-            request.FILES
-        )
-        if upload_form.is_valid():
-            if upload_form.cleaned_data['database'] == 'pcidatabase.com':
-                function = utils.pcidatabase__parse_graphics_devices_iterable
-            else:
-                function = utils.pci_ids__parse_graphics_devices_iterable
-
-            payload = list(function(upload_form.cleaned_data['file']))
-            api = GraphicsDevices()
-            result = api.post(data=payload)
-            messages.success(
-                request,
-                'Graphics device CSV upload successfully saved.'
-            )
-            return redirect('manage:graphics_devices')
-
-    elif request.method == 'POST':
-        form = forms.GraphicsDeviceForm(request.POST)
-        if form.is_valid():
-            payload = [{
-                'vendor_hex': form.cleaned_data['vendor_hex'],
-                'adapter_hex': form.cleaned_data['adapter_hex'],
-                'vendor_name': form.cleaned_data['vendor_name'],
-                'adapter_name': form.cleaned_data['adapter_name'],
-            }]
-            api = GraphicsDevices()
-            result = api.post(data=payload)
-            if result:
-                messages.success(
-                    request,
-                    'Graphics device saved.'
-                )
-            return redirect('manage:graphics_devices')
-
-    context['page_title'] = "Graphics Devices"
-    context['form'] = form
-    context['upload_form'] = upload_form
-    return render(request, 'manage/graphics_devices.html', context)
-
-
-@json_view
-@superuser_required
-def graphics_devices_lookup(request):
-    form = forms.GraphicsDeviceLookupForm(request.GET)
-    if form.is_valid():
-        vendor_hex = form.cleaned_data['vendor_hex']
-        adapter_hex = form.cleaned_data['adapter_hex']
-        api = GraphicsDevices()
-        result = api.get(vendor_hex=vendor_hex, adapter_hex=adapter_hex)
-        return result
-    else:
-        return http.HttpResponseBadRequest(str(form.errors))
